update/ remove excess console.log
