+++ modules/LandingPage/data.tsx
@@ -3,6 +3,17 @@ import { CommentDataType } from "@/constants/types/commentDataType";
 import { GroupDataType } from "@/constants/types/groupDataType";
 import { CitiesDataType } from "@/constants/globalTypes";
 
+// page
+
+export type HomeProps = {
+  citiesData: CitiesDataType;
+  commentsData: CommentDataType[];
+  nearbyGroupsData: GroupDataType[];
+  remainingGroupsData: GroupDataType[];
+  preferenceData: GroupDataType[];
+  nearbyStoresData: StoreDataType[];
+};
+
 // CardsSectionProps
 export type CardsSectionProps = {
   layout: "swipe" | "block";
