+++ pages/index.tsx
@@ -1,14 +1,10 @@
-import { createContext, useContext, useEffect, useState } from "react";
+import { createContext } from "react";
 import { GetServerSidePropsContext } from "next";
 import { Inter } from "next/font/google";
 import fetchApi, { apiParamsType } from "@/common/helpers/fetchApi";
 import apiPaths from "@/constants/apiPaths";
-import Layout from "@/common/components/Layout";
 import LandingPage from "@/modules/LandingPage";
-import { CommentDataType } from "@/constants/types/commentDataType";
-import { CitiesDataType } from "@/constants/globalTypes";
-import { GroupDataType } from "@/constants/types/groupDataType";
-import { StoreDataType } from "@/constants/types/storeDataType";
+import Layout from "@/common/components/Layout";
 import {
   defaultGroupsSearchKey,
   getSearchGroups,
@@ -18,6 +14,7 @@ import {
   getSearchStores,
 } from "@/common/helpers/getSearchApi/getSearchStores";
 import {
+  HomeProps,
   defaultCitiesData,
   defaultCommentsData,
   defaultGroupsData,
@@ -56,18 +53,7 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     const commentsRes = await fetchApi(commentsApiParams);
     commentsData = commentsRes?.data ?? [];
 
-    // 搜尋揪團的 APIparams
-    const searchGroupApiParams: apiParamsType = {
-      apiPath: apiPaths["search-group"],
-      method: "POST",
-    };
-
-    const searchStoreApiParams: apiParamsType = {
-      apiPath: apiPaths["search-store"],
-      method: "POST",
-    };
-
-    // 有登入用戶的喜好設定 => 要問洛陽 gameType 的篩選?
+    // 有登入用戶的喜好設定
     const preferenceApiParams: apiParamsType = {
       apiPath: apiPaths["get-preference-group"],
       method: "POST",
@@ -113,8 +99,6 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     };
 
     nearbyStoresData = await getSearchStores(searchLocationStores);
-
-    //---
   } catch (error) {
     console.error(error);
   }
@@ -130,24 +114,6 @@ export async function getServerSideProps(context: GetServerSidePropsContext) {
     },
   };
 }
-
-// 搜尋的地點內容：城市
-// 搜尋的日期內容
-// 用戶資料（如果有登入）
-// 就差你一個成團 => 可以直接打
-// 你附近的揪團 => IP位置
-// 你可能有興趣的揪團 => 先取得用戶資料、以遊戲喜好搜尋（有搜喜好地區嗎？）
-// 你附近的店家 => IP位置
-// 評價
-
-export type HomeProps = {
-  citiesData: CitiesDataType;
-  commentsData: CommentDataType[];
-  nearbyGroupsData: GroupDataType[];
-  remainingGroupsData: GroupDataType[];
-  preferenceData: GroupDataType[];
-  nearbyStoresData: StoreDataType[];
-};
 
 export const defaultDataContext = {
   citiesData: defaultCitiesData,
@@ -168,10 +134,6 @@ export default function Home({
   preferenceData,
   nearbyStoresData,
 }: HomeProps) {
-  // const authContext = useContext(AuthContext);
-  // const { authData } = authContext;
-  // const logout = useLogout();
-
   const dataSet = {
     citiesData: citiesData,
     commentsData: commentsData,
@@ -180,8 +142,6 @@ export default function Home({
     preferenceData: preferenceData,
     nearbyStoresData: nearbyStoresData,
   };
-
-  console.log("data", dataSet);
 
   return (
     <>
